nif: add simple_log
Explore NIFs, rustler, and jj-lib, by implementing a simple function to show a
simple log, from the working copy commit, up to 10 ancestors. This function can
be invoked in Elixir by going:

    workspace = Jj.Native.get_workspace("/path/to/jj/repo")
    Jj.Native.simple_log(workspace)

Create a `get_workspace` function, which takes in a path and returns a reference
to a jj_lib::workspace::Workspace object. The initial idea is to keep this
reference on the elixir side, and pass it in whenever we want to invoke a jj
command on this workspace.