nif/web: add logging from revset
On mount, the page renders the log based on `simple_log` (we will remove this in
the future), and also loads the default revset (under `revsets.log` in settings)
in a input field. Hitting enter in the field triggers the new `log` function.

Right now the log output shows up as a simple list of commits. We'll add better
visualisations later.

Also, the code is pretty scrappy, we'll leave the refactoring for later.