refactor(nif): break up lib.rs, remove simple_log
Rename Workspace -> WorkspaceSettings since we are passing more than just the
Workspace into Elixir-land.

Wrap WorkspaceSettings into a larger WorkspaceData. This contain things that do
not implement `Send` and `Sync`, so they can't be registered as Resources.

Also register a function to get the RevsetParseContext because it borrows so
many things, and I can't figure out a way to simply "cache" a single instance of
it inside WorkspaceData for reuse, so simply construct a new one, rather than
reusing a cached one. I stole this idea from GG.