fix(web): handle missing `revset.log` in config file
When `revset.log` is missing from user's config file, the web app fails to load
valid `jj` repositories. To fix this, we fallback to a default revset expression
when we're unable to find `revset.log` in the user's config.

However, it is worthy to note the default log revset introduced in this PR
differs from `jj-lib`'s default fallback. The revset expression we are using
shows the latest 10 commits relative to the current working copy 
(`ancestors(@, 10)`).